+++ bukkit/src/main/java/com/github/games647/fastlogin/bukkit/listener/protocolsupport/ProtocolSupportListener.java
@@ -70,6 +70,10 @@ public class ProtocolSupportListener implements Listener {
                         PlayerProfile uuidProfile = plugin.getCore().getStorage().loadProfile(premiumUUID);
                         if (uuidProfile != null) {
                             plugin.getLogger().log(Level.FINER, "Player {0} changed it's username", premiumUUID);
+
+                            //update the username to the new one in the database
+                            uuidProfile.setPlayerName(username);
+
                             startPremiumSession(username, loginStartEvent, false, uuidProfile);
                             return;
                         }